invert XRE autoconf logic
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@639 941e2c2b-deac-454f-805a-451daa25f33c

+++ libhfst/src/implementations/xre_utils.cc
@@ -46,7 +46,7 @@ std::map<std::string,hfst::HfstTransducer> definitions;
 char* startptr;
 hfst::HfstTransducer* last_compiled;
 hfst::ImplementationType format;
-unsigned int len;
+size_t len;
 
 int*
 get_n_to_k(const char* s)
